[add] 날짜별 관광지 가져오기 추가

+++ yolo/src/main/java/com/yolo/dto/TripDto.java
@@ -11,8 +11,8 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TripDto {
-	private Long contentId;
-	private Long conteTypeId;
+	private int contentId;
+	private int conteTypeId;
 	private String title;
 	private String address;
 	private String imageUrl;
